CCM-30: Added correct consent category to video iframes
Services like Youtube and Vimeo provide by default iframes that track
the users with cookies. They provide a no-cookie domain but that is not
wat is used by default.

To avoid loading this iframes and the cookies that come with them,
Cookiebot removes the iframe content until the visitor has allowed
marketing cookies.

This fix adds the proper iframe consent category to video_embed_fields
if privacy is enabled.